requirements.txt
asgiref==3.8.1
bleach==6.1.0
blinker==1.7.0
boto3==1.35.16
botocore==1.35.39
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
crispy-bootstrap4==2024.10
crispy-bootstrap5==0.7
cryptography==42.0.5
DateTime==5.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==5.1
django-allauth==0.54.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.5
django-extensions==4.1
django-storages==1.14.4
django-summernote==0.8.20.0
django_csp==3.8
Flask==3.0.2
Flask-Cors==5.0.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-spreadsheet==0.0.6
googleapis-common-protos==1.62.0
greenlet==3.0.3
gspread==5.12.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
np==1.0.2
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==25.0
pandas==2.1.4
pillow==10.4.0
protobuf==4.25.1
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pyfiglet==1.0.2
PyJWT==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.3
setuptools==69.0.2
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.5.1
stripe==10.10.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.7.0
zope.interface==6.1
